refactor: 신고 모달 공통 컴포넌트 추출 및 코드 개선(#379)
- formatDate, Field, DeleteConfirmDialog를 common/으로 추출하여 중복 제거
- report.images[idx] → report.images?.[idx] optional chaining 적용
- UserReportDetailModal의 formatDate를 toLocaleDateString에서 수동 포맷으로 통일
