+++ src/features/admin/components/common/DeleteConfirmDialog.tsx
@@ -1,0 +1,65 @@
+'use client'
+
+import { useEffect, useRef } from 'react'
+
+interface DeleteConfirmDialogProps {
+  isOpen: boolean
+  title: string
+  description: string
+  confirmLabel?: string
+  onConfirm: () => void
+  onCancel: () => void
+}
+
+export default function DeleteConfirmDialog({
+  isOpen,
+  title,
+  description,
+  confirmLabel = '삭제',
+  onConfirm,
+  onCancel,
+}: DeleteConfirmDialogProps) {
+  const confirmRef = useRef<HTMLDialogElement>(null)
+
+  useEffect(() => {
+    const dialog = confirmRef.current
+    if (isOpen && !dialog?.open) {
+      dialog?.showModal()
+    } else if (!isOpen && dialog?.open) {
+      dialog?.close()
+    }
+  }, [isOpen])
+
+  return (
+    <dialog
+      ref={confirmRef}
+      className="m-auto w-full max-w-fit open:flex flex-col gap-4 rounded-xl bg-white p-6 shadow-2xl backdrop:bg-gray-900/50"
+      onClick={(e) => {
+        if (e.target === confirmRef.current) onCancel()
+      }}
+      onClose={(e) => {
+        e.stopPropagation()
+        onCancel()
+      }}
+    >
+      <h4 className="text-lg font-semibold text-gray-900">{title}</h4>
+      <p className="text-sm leading-relaxed text-gray-500">{description}</p>
+      <div className="flex justify-end gap-3">
+        <button
+          type="button"
+          onClick={onCancel}
+          className="cursor-pointer rounded-lg border border-gray-300 bg-white px-4 py-2 text-sm font-medium text-gray-700 hover:bg-gray-50"
+        >
+          취소
+        </button>
+        <button
+          type="button"
+          onClick={onConfirm}
+          className="cursor-pointer rounded-lg bg-red-600 px-4 py-2 text-sm font-medium text-white hover:bg-red-700"
+        >
+          {confirmLabel}
+        </button>
+      </div>
+    </dialog>
+  )
+}
