+++ src/features/admin/components/common/formatDate.ts
@@ -1,0 +1,7 @@
+export function formatDate(iso: string) {
+  const d = new Date(iso)
+  const yyyy = d.getFullYear()
+  const MM = String(d.getMonth() + 1).padStart(2, '0')
+  const dd = String(d.getDate()).padStart(2, '0')
+  return `${yyyy}-${MM}-${dd}`
+}
